renaming links pattern
